.requirements.txt
Django~=5.1.5
psycopg2-binary
asyncpg
python-dotenv
celery[redis]
redis
django-redis
python-telegram-bot[job-queue]~=21.10
python-telegram-bot-calendar
aioredis
channels
channels_redis
gunicorn
uvicorn
django-cors-headers
django-admin-wizard
django-ninja
django-ninja-extra
django-ninja-jwt
